Change threading to multiprocessing
Multiprocessing seems to be more optimized as in being an actual
cpu-threading. I think it's better not to waste cpu time on a
pseudo-threading but rather use more of it to make the impact from the
application negligible.
May revent the change if I would feel I need to, so far it makes sense.