Add more interactivity, fix the subprocess issue
So the issue was that the windows shell sometimes requires the extensions
to be written explicitly. Just solved the case for Windows 10 when it
wouldn't detect livestreamer unless it was explicitly saying that it's an
.exe
Added more interactivity to shake up the awkward silence while the
processes are preparing. Should do the trick.